+++ Scripts/Extensions/AudioGroupExtension.cs
@@ -15,12 +15,12 @@ namespace TinyServices.Audio.Extensions {
             return group.GetConfigs()[type].Play(position);
         }
         
-        public static AudioSource Play<T>(this AudioGroup<T> group, T type, float limit) where T : Enum {
+        public static AudioSource Play<T>(this AudioGroup<T> group, T type, int limit) where T : Enum {
             string key = $"{type.GetType().Name}_{type}";
             return group.GetConfigs()[type].Play(key, limit);
         }
         
-        public static AudioSource Play<T>(this AudioGroup<T> group, T type, Vector3 position, float limit) where T : Enum {
+        public static AudioSource Play<T>(this AudioGroup<T> group, T type, Vector3 position, int limit) where T : Enum {
             string key = $"{type.GetType().Name}_{type}";
             return group.GetConfigs()[type].Play(position, key, limit);
         }
